02/requirements.txt
anthropic>=0.25.0
python-telegram-bot>=20.0
flask>=2.0
schedule>=1.2
